Exceptions to let the project run outside of Pis
